chore: move `FilterControls` interface
