+++ website-frontend/src/routes/people/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	/** @type {import('./$types').PageData} */
-	import type { FilterControls } from '$lib/components/filter/filter_controls';
+	import type { FilterControls } from '$lib/types/filter_controls';
 	import Banner from '$lib/components/people/Banner.svelte';
 	import FilterBar from '$lib/components/filter/FilterBar.svelte';
 	import LoadMore from '$lib/components/load_more/LoadMore.svelte';
